+++ ElectronicObserver.KancolleApi.Types/Models/APIRaigekiClass.cs
@@ -30,7 +30,7 @@ public class ApiRaigekiClass
 	[JsonPropertyName("api_fdam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiFdam { get; set; } = new();
+	public List<double> ApiFdam { get; set; } = new();
 
 	[JsonPropertyName("api_frai")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
